requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langgraph>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.5.0
typer[all]>=0.9.0
rich>=13.6.0
huggingface-hub>=0.20.0
transformers>=4.38.0
sentence-transformers>=2.3.1
duckduckgo-search>=4.1.1
beautifulsoup4>=4.12.2
requests>=2.31.0 